+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [6.1.0] - 2022-02-02
+## [6.2.0] - 2022-03-07
 ### Added
 - Ability to create a cohort of patients by browsing and selecting records
   for non-dbGap FHIR servers.
@@ -14,10 +14,24 @@ This project follows [Semantic Versioning](http://semver.org/).
   searching for patients.
 - An exception when trying to pull data for an empty patient cohort.
 
-## [6.0.8] - 2022-01-26
+## [6.1.0] - 2022-02-23
+### Changed
+- Hid an unnecessary option to view the full list of research studies from
+  the select records step.
+- Added aborting of HTTP requests if we no longer need a response to them.
+
+## [6.0.10] - 2022-02-13
 ### Added
 - An input field for the maximum number of recent Observations per Patient
   to check when no code is specified in the criteria.
+
+## [6.0.9] - 2022-02-06
+### Changed
+- Temporarily disabled the use of the _security URL parameter.
+
+## [6.0.8] - 2022-02-02
+### Changed
+- Enabled use of RAS for dbGap by default.
 
 ## [6.0.7] - 2022-01-26
 ### Changed